Simplified a lot the sample
