+++ samples/teamsfx-app/package.json
@@ -22,11 +22,13 @@
   },
   "devDependencies": {
     "@types/node": "^14.0.0",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@types/react-router-dom": "^5.3.3",
     "env-cmd": "^10.1.0",
     "typescript": "^4.1.2"
+  },
+  "peerDependencies": {
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/react-router-dom": "^5.3.3"
   },
   "scripts": {
     "dev:teamsfx": "env-cmd --silent -f .localConfigs npm run start",
